Add "Component versions in recent OpenVox releases" reference page
OpenVox ships its components on independent version lines (openvox-agent,
openvox-server, openvoxdb, and OpenBolt all release separately with no shared
version), so there is no single bundled version that answers "what's in this
release?". Add a dedicated reference page with one generated table per component,
each keyed by that component's own release.

- New generators (mirroring AgentReleaseTable) resolve bundled versions from
  upstream component pins: ServerReleaseTable (server -> JRuby via the
  jruby-utils/jruby-deps chain), OpenvoxdbReleaseTable (release tags), and
  OpenboltReleaseTable (openbolt -> Ruby/OpenSSL/r10k). Unresolvable older tags
  skip with a warning; an empty resolve refuses to overwrite committed data.
- Rake tasks references:server_versions, references:openvoxdb_versions,
  references:openbolt_versions, plus an umbrella references:component_versions;
  wired into build.yaml with committed-data fallback.
- Multi-series ready: the per-OpenVox-series tables (agent/server/openvoxdb)
  write into _data/<table>/<nav_key>.yml and the page renders its own series via
  site.data.<table>[page.nav], so a future 9.x collection gets its own data file
  with no collision (validated with a throwaway 9.x cutover). OpenBolt is
  independent of the OpenVox major and stays shared in a single data file.
- Fix a pre-existing bug: the agent table listed r10k, but r10k is bundled by
  openbolt-runtime, not the agent runtime (agent-runtime-8.x), so it does not
  ship in openvox-agent. Drop r10k from AgentReleaseTable and the agent table,
  and surface it in the OpenBolt table where it is actually bundled. Java and
  PostgreSQL are supported-version requirements, not pins, so they are shown as
  hand-maintained notes; openvoxdb-termini tracks openvoxdb and is noted, not
  tabled.
- Move the agent release-contents table off about_agent.md (now a pointer) to
  avoid duplication; add nav entry and repoint the collection index link.

Part of #333
