+++ _data/openvoxdb_release_contents/openvox_8x.yml
@@ -1,0 +1,5 @@
+---
+- release: 8.14.0
+- release: 8.13.0
+- release: 8.12.1
+- release: 8.12.0
