+++ _data/server_release_contents/openvox_8x.yml
@@ -1,0 +1,9 @@
+---
+- release: 8.14.0
+  jruby: 9.4.12.1
+- release: 8.13.0
+  jruby: 9.4.12.1
+- release: 8.12.1
+  jruby: 9.4.12.1
+- release: 8.12.0
+  jruby: 9.4.12.1
